% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_trace_regression.R
\name{bic.quantile_trace_regression}
\alias{bic.quantile_trace_regression}
\title{cv Quantile Trace Regression}
\usage{
bic.quantile_trace_regression(
  X,
  y,
  tau,
  nlambda,
  B_init,
  B,
  gamma = 0.1,
  MAXIT = 10000L,
  eps = 0.01,
  quiet = T
)
}
\arguments{
\item{gamma}{the step size}
}
\description{
cv Quantile Trace Regression
}
