% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_regression.R
\name{bic.trace_regression}
\alias{bic.trace_regression}
\title{cv Trace Regression}
\usage{
bic.trace_regression(
  X,
  y,
  nlambda,
  B_init,
  B,
  gamma = 0.1,
  MAXIT = 10000L,
  eps = 0.01,
  quiet = T
)
}
\arguments{
\item{gamma}{the step size}
}
\description{
cv Trace Regression
}
