% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_trace_regression.R
\name{quantile_trace_regression}
\alias{quantile_trace_regression}
\title{Quantile Trace Regression}
\usage{
quantile_trace_regression(
  X,
  y,
  tau,
  lambda,
  B_init,
  B,
  gamma = 0.1,
  MAXIT = 10000L,
  eps = 0.01,
  quiet = T
)
}
\arguments{
\item{gamma}{the step size}
}
\description{
Quantile Trace Regression
}
